seperate files for editor and constants
