+++ src/constants.js
@@ -1,0 +1,15 @@
+// Supported feature flags
+export const SUPPORT_PAUSE = 1;
+export const SUPPORT_SEEK = 2;
+export const SUPPORT_VOLUME_SET = 4;
+export const SUPPORT_VOLUME_MUTE = 8;
+export const SUPPORT_PREVIOUS_TRACK = 16;
+export const SUPPORT_NEXT_TRACK = 32;
+export const SUPPORT_TURN_ON = 128;
+export const SUPPORT_TURN_OFF = 256;
+export const SUPPORT_PLAY_MEDIA = 512;
+export const SUPPORT_STOP = 4096;
+export const SUPPORT_PLAY = 16384;
+export const SUPPORT_SHUFFLE = 32768;
+export const SUPPORT_GROUPING = 524288;  
+export const SUPPORT_REPEAT_SET = 262144;
